5.8

+++ src/init.js
@@ -4,7 +4,7 @@ import config from '~/config.js'
 
 let initVer=()=>{
 	let hxxver = localStorage.getItem('hxxver'), ver= parseFloat(config.versions);
-	if(!hxxver || (hxxver && parseInt(hxxver)<ver)){
+	if(!hxxver || (hxxver && parseFloat(hxxver)<ver)){
 		localStorage.clear()
 		localStorage.setItem('hxxver', ver)
 	}
